refactor: Move process out of Map and Parallel into the base and we're done.
