+++ lib/state-machines/state-types/Map.js
@@ -1,7 +1,5 @@
-'use strict'
 const ParentBaseState = require('./Parent-base-state')
 const jp = require('jsonpath')
-const States = require('./errors')
 
 class Map extends ParentBaseState {
   constructor (stateName, stateMachine, stateDefinition, options) {
@@ -18,27 +16,6 @@ class Map extends ParentBaseState {
       this.makeChildExecution(this.iterator, item, index, parentExecutionName)
     )
   } // buildBranchExecutions
-
-  async process (executionDescription, input) {
-    const parentExecutionName = executionDescription.executionName
-    await this.dao.checkpoint(executionDescription)
-
-    const branchExecutions = this.buildBranchExecutions(input, parentExecutionName)
-
-    Promise
-      .all(branchExecutions)
-      .then(() => this.handleChildExecutionComplete(
-        parentExecutionName,
-        this.stateMachineName,
-        this.name
-      ))
-      .catch(err => {
-        this.processTaskFailure(
-          States.BranchFailed.error(err.toString()),
-          parentExecutionName.executionName
-        )
-      })
-  }
-}
+} // class Map
 
 module.exports = Map
