+++ lib/state-machines/state-types/Parallel.js
@@ -1,6 +1,4 @@
-'use strict'
 const ParentBaseState = require('./Parent-base-state')
-const States = require('./errors')
 
 class Parallel extends ParentBaseState {
   constructor (stateName, stateMachine, stateDefinition, options) {
@@ -22,27 +20,6 @@ class Parallel extends ParentBaseState {
       this.makeChildExecution(stateMachineName, input, index, parentExecutionName)
     )
   } // buildBranchExecutions
-
-  async process (executionDescription, input) {
-    const parentExecutionName = executionDescription.executionName
-    await this.dao.checkpoint(executionDescription)
-
-    const branchExecutions = this.buildBranchExecutions(input, parentExecutionName)
-
-    Promise
-      .all(branchExecutions)
-      .then(() => this.handleChildExecutionComplete(
-        parentExecutionName,
-        this.stateMachineName,
-        this.name
-      ))
-      .catch(err => {
-        this.processTaskFailure(
-          States.BranchFailed.error(err.toString()),
-          parentExecutionName.executionName
-        )
-      })
-  } // process
 } // class Parallel
 
 module.exports = Parallel
