+++ lib/state-machines/state-types/Parent-base-state.js
@@ -1,5 +1,6 @@
 const BaseStateType = require('./Base-state')
 const cloneDeep = require('lodash/cloneDeep')
+const States = require('./errors')
 
 class ParentBaseState extends BaseStateType {
   constructor (stateName, stateMachine, stateDefinition, options) {
@@ -28,11 +29,28 @@ class ParentBaseState extends BaseStateType {
       },
       this.options
     )
-  }
+  } // makeChildExecution
 
   async process (executionDescription, input) {
+    const parentExecutionName = executionDescription.executionName
+    await this.dao.checkpoint(executionDescription)
 
-  }
+    const branchExecutions = this.buildBranchExecutions(input, parentExecutionName)
+
+    Promise
+      .all(branchExecutions)
+      .then(() => this.handleChildExecutionComplete(
+        parentExecutionName,
+        this.stateMachineName,
+        this.name
+      ))
+      .catch(err => {
+        this.processTaskFailure(
+          States.BranchFailed.error(err.toString()),
+          parentExecutionName.executionName
+        )
+      })
+  } // process
 } // class ParentBaseState
 
 module.exports = ParentBaseState
